gpt5.2-deepgramflux-elevenflashv2.5-livekit/pyproject.toml
[project]
name = "gpt5.2-deepgramflux-elevenflashv2.5-livekit"
version = "0.1.0"
description = "Voice agent using LiveKit Agents with Deepgram Flux STT, GPT-5.2-mini LLM, ElevenLabs TTS, and Plivo SIP telephony"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "livekit-agents>=1.5.0",
    "livekit-plugins-deepgram>=1.5.0",
    "livekit-plugins-openai>=1.5.0",
    "livekit-plugins-elevenlabs>=1.5.0",
    "livekit-plugins-silero>=1.5.0",
    "livekit-plugins-noise-cancellation>=0.2.0",
    "livekit-api>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "plivo>=4.59.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "phonenumbers>=8.13.0",
    "numpy>=1.24.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "websockets>=12.0",
    "faster-whisper>=1.2.1",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP", "SIM", "RUF"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
